Set up first test for the blog search page
