+++ network-api/networkapi/wagtailpages/tests/research_hub/base.py
@@ -4,30 +4,14 @@ from wagtail_localize import synctree
 
 from networkapi.wagtailpages.factory import homepage as home_factory
 from networkapi.wagtailpages.factory import research_hub as research_factory
+from networkapi.wagtailpages.tests import base as test_base
 
 
-class ResearchHubTestCase(test.TestCase):
+class ResearchHubTestCase(test_base.WagtailpagesTestCase):
     @classmethod
     def setUpTestData(cls):
-        cls._setup_homepage()
+        super().setUpTestData()
         cls._setup_research_hub_structure(homepage=cls.homepage)
-        cls._setup_locales()
-
-    @classmethod
-    def _setup_homepage(cls):
-        root = wagtail_models.Page.get_first_root_node()
-        if not root:
-            raise ValueError('A root page should exist. Something is off.')
-        cls.homepage = home_factory.WagtailHomepageFactory(parent=root)
-
-        sites = wagtail_models.Site.objects.all()
-        if sites.count() != 1:
-            raise ValueError('There should be exactly one site. Something is off.')
-        cls.site = sites.first()
-
-        cls.site.root_page = cls.homepage
-        cls.site.clean()
-        cls.site.save()
 
     @classmethod
     def _setup_research_hub_structure(cls, homepage):
@@ -42,19 +26,5 @@ class ResearchHubTestCase(test.TestCase):
             title='Authors',
         )
 
-    @classmethod
-    def _setup_locales(cls):
-        cls.default_locale = wagtail_models.Locale.get_default()
-        cls.fr_locale, _ = wagtail_models.Locale.objects.get_or_create(
-            language_code='fr'
-        )
-        assert cls.fr_locale != cls.default_locale
-
     def setUp(self):
         self.synchronize_tree()
-
-    def synchronize_tree(self):
-        synctree.synchronize_tree(
-            source_locale=self.default_locale,
-            target_locale=self.fr_locale
-        )
